+++ lib/unifig/config.rb
@@ -5,7 +5,7 @@ module Unifig
   class Config
     # @raise [MissingConfigError] - No config section was provided.
     def initialize(config, env: nil)
-      raise MissingConfigError, 'You must provide a configuration' unless config
+      raise MissingConfigError, 'no configuration provided' unless config
 
       @env_config = config.slice(:providers)
       @env = env